+++ solid/frame/mprpc/test/test_relay_close_response.cpp
@@ -55,7 +55,7 @@ using MessageIdVectorT = std::deque<MessageIdT>;
 
 std::string            pattern;
 const size_t           initarraysize = sizeof(initarray) / sizeof(InitStub);
-std::atomic<size_t>    crtwriteidx(0);
+std::atomic<uint32_t>  crtwriteidx(0);
 std::atomic<size_t>    writecount(0);
 std::atomic<size_t>    created_count(0);
 std::atomic<size_t>    canceled_count(0);
